+++ CHANGELOG.md
@@ -1,11 +1,14 @@
-## [2.0.3] - [June 2, 2022]
+## [2.0.3] - [June 13, 2022]
 * Updated documentation
 * Fixed display of web support on pub.dev
 * Fixed scrollbar on wide screens
+* Fixed 'trailing' when used 'SettingsTile.switchTile'
+* Fixed color scheme on web version
+* Fixed default vertical padding on web version
 
 ## [2.0.2] - [January 24, 2022]
 * Enabled parameter for the ListTile widget
-* Trailing parameter for the ListTile widget
+* Trailing parameter for the ListTile widget 
 * Fixed minor bugs
 
 ## [2.0.1] - [January 9, 2022]